+++ unstructured/metrics/table/table_eval.py
@@ -41,6 +41,7 @@ class TableEvaluation:
     """Class representing a gathered table metrics."""
 
     total_tables: int
+    total_predicted_tables: int
     table_level_acc: float
     table_detection_recall: float
     table_detection_precision: float
@@ -247,6 +248,7 @@ class TableEvalProcessor:
             table_acc = 1 if not is_table_predicted else 0
             return TableEvaluation(
                 total_tables=0,
+                total_predicted_tables=len(predicted_table_data),
                 table_level_acc=table_acc,
                 table_detection_recall=score,
                 table_detection_precision=score,
@@ -259,6 +261,7 @@ class TableEvalProcessor:
         if is_table_in_gt and not is_table_predicted:
             return TableEvaluation(
                 total_tables=len(ground_truth_table_data),
+                total_predicted_tables=0,
                 table_level_acc=0,
                 table_detection_recall=0,
                 table_detection_precision=0,
@@ -294,6 +297,7 @@ class TableEvalProcessor:
 
             evaluation = TableEvaluation(
                 total_tables=len(ground_truth_table_data),
+                total_predicted_tables=len(predicted_table_data),
                 table_level_acc=predicted_table_acc,
                 table_detection_recall=table_detection_recall,
                 table_detection_precision=table_detection_precision,